Sponsorに従属する関数をsponsorType.tsxへ移動
